Make the Python impl handle setting __class__ and __dict__ on ghosts like C.
By activating the object.

Add tests for this.

In the process of adding tests, I discovered the mock pickle cache used by the Python tests
was returning the wrong thing from new_ghost (not a ghost), so I fixed this and added some extra
tests and assertions to be sure the behaviour matches with the real and C caches.

Fixes #155